GPG/CCID OpenPGP smartcard + BLE/WiFi coexistence (0.5.1)
CCID OpenPGP 3.4 implementation
- Sign / verify / encrypt / decrypt / SSH auth end-to-end with GnuPG
- T=1 block protocol, ISO 7816-4 command chaining, GET RESPONSE
- TROPIC01 ECDSA/EdDSA in hardware (SIG/AUT slots)
- ECDH on ESP32 (DEC privkey AES-256-GCM-wrapped in R-Mem 502, chip-bound
  HKDF storage key, plaintext zeroized immediately after use)
- INTERNAL_AUTHENTICATE, RESET_RETRY_COUNTER (RC + admin), TERMINATE_DF
  and ACTIVATE_FILE, MSE, writable algorithm attributes (C1/C2/C3)
- Dual-store unification: device UI reads OpenPGP card-application state
  so CCID-generated keys are visible on the badge

Security audit pass
- Serial command echo no longer logs PINs / passwords / TOTP seeds
- GPG_RESET serial command requires a 30-second one-shot confirmation token
- Resetting Code stored as iterated salted SHA-256, never plaintext
- OpenPGP NVS state ECDSA-signed by the attestation slot; tampered blobs
  trigger silent reinit
- CDC log default level gated by DEBUG_MODE; INFO/DEBUG unlock on auth
- PIN retry counter pre-decrement is power-cycle safe; redundant R-Memory
  writes skipped via in-RAM mirror
- AES-GCM nonces use getRandomStrict (no software-RNG fallback)

BLE + WiFi simultaneous coexistence
- NimBLE host pools moved to PSRAM (EXTERNAL allocation mode)
- lwIP allocations preferentially in PSRAM
- BT controller MAX_ACT and ADV duplicate filter trimmed
- NimBLE MAX_CONNECTIONS / CCCDS / MSYS / ACL_BUF rebalanced
- Task stacks tuned (main, usbd, fido2, epd_render, keypad) to measured
  worst-case watermarks under crypto stress
- ~60 KB internal DRAM freed; esp_wifi_init no longer hits NO_MEM with
  BLE already running

Hardware and diagnostics
- meminfo serial command shows DRAM, DMA-capable and PSRAM separately,
  plus per-task stack watermarks with DRAM/PSRAM location
- HardwareInfo view shows DRAM + DMA-capable + PSRAM
- Custom partitions: 256 KB NVS, factory app slot, drop SPIFFS + OTA-B
- meminfo and HardwareInfo expose largest-free-block + min-ever-free
  to make fragmentation visible

Device UI
- GPG menu hides Generate Keys when keys exist
- Status shows "FP:" instead of "Fingerprint:" so it fits in one line
- Export QR shows user-id, curve and last-4 fingerprint bytes on the
  right side; only N closes the QR view
- Export reads SIG pubkey directly from TROPIC01 when device-side metadata
  is empty (CCID-generated keys)
- Serial AUTH without PIN logs out; wrong PIN drops the active session

Project housekeeping
- docs/GPG.md rewritten (hardware/software split, threat model, end-to-end
  SSH bootstrap, troubleshooting)
- docs/README.md updated to reflect Working status with UI WIP note
- docs/SERIAL_COMMANDS.md updated for GPG_RESET two-step
- 47 MB of tests/, audit/, plan*.md removed from the tree
- mod_homeassistant skeleton registered as headers-only stub
- mod_vcard wizard view scaffolded
- partitions.csv pinned (256 KB NVS, factory + coredump)
- Bump to 0.5.1

+++ components/mod_gpg/src/openpgp/apdu.cpp
@@ -42,7 +42,7 @@ bool apdu_parse(const uint8_t *raw, size_t raw_len, apdu_t *apdu) {
 
         // Extended Lc (3 bytes: 0x00 + 2 bytes)
         if (pos + 3 <= raw_len) {
-            apdu->lc = (raw[pos + 1] << 8) | raw[pos + 2];
+            apdu->lc = static_cast<uint16_t>((raw[pos + 1] << 8) | raw[pos + 2]);
             pos += 3;
         }
     } else {
@@ -75,7 +75,7 @@ bool apdu_parse(const uint8_t *raw, size_t raw_len, apdu_t *apdu) {
         if (apdu->extended) {
             // Extended Le (2 bytes)
             if (pos + 2 <= raw_len) {
-                uint16_t le_val = (raw[pos] << 8) | raw[pos + 1];
+                uint16_t le_val = static_cast<uint16_t>((raw[pos] << 8) | raw[pos + 1]);
                 apdu->le = (le_val == 0) ? 65536 : le_val;
             }
         } else {
